Renamed test/unit/helper.rb to test/helper.rb
This fixes using `m` and other tooling due to bad requires /
non-standard pathing.
