+++ Rakefile
@@ -22,7 +22,7 @@ task lint: [:rubocop]
 task default: [:test, :lint]
 
 Rake::TestTask.new do |t|
-  t.libs << 'test/unit'
+  t.libs << 'test'
   t.pattern = 'test/unit/**/*_test.rb'
   t.warning = true
   t.verbose = true
